+++ lib/.storybook/main.js
@@ -1,4 +1,5 @@
 const { vanillaExtractPlugin } = require("@vanilla-extract/vite-plugin");
+
 module.exports = {
   stories: ["../src/**/*.stories.mdx", "../src/**/*.stories.@(js|jsx|ts|tsx)"],
   addons: ["@storybook/addon-links", "@storybook/addon-essentials"],
@@ -7,7 +8,8 @@ module.exports = {
     builder: "storybook-builder-vite",
   },
   async viteFinal(config) {
-    config.plugins.push(vanillaExtractPlugin());
+    config.plugins = [...config.plugins, vanillaExtractPlugin()];
+
     return config;
   },
 };
