--- lib/src/components/reset/Button/index.ts
@@ -1,1 +1,0 @@
-export * from "./Button";
