+++ lib/src/index.ts
@@ -1,6 +1,3 @@
 export * from "./components/Button";
+export * from "./components/Card";
 export * from "./components/reset";
-
-import { Button as ButtonStyles } from "./components/Button";
-
-export { ButtonStyles };
