+++ lib/vite.config.ts
@@ -1,12 +1,12 @@
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
+
 import { vanillaExtractPlugin } from "@vanilla-extract/vite-plugin";
 import { resolve } from "path";
 import tsconfigPaths from "vite-tsconfig-paths";
 import dts from "vite-plugin-dts";
 
 import pkg from "./package.json";
-
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [
@@ -36,7 +36,7 @@ export default defineConfig({
     lib: {
       entry: "src/index.ts",
       fileName: "index",
-      formats: ["cjs", "es"],
+      formats: ["es"],
     },
     rollupOptions: {
       external: Object.keys(pkg.peerDependencies),
