+++ package.json
@@ -1,4 +1,5 @@
 {
+  "private": true,
   "name": "@hover-design/react",
   "version": "1.0.0",
   "description": "Hover Component library created by bunch of ants from antstack",
@@ -8,9 +9,13 @@
     "type": "git",
     "url": "git+https://github.com/antstackio/hover-design.git"
   },
-  "dependencies": {},
-  "devDependencies": {},
+  "devDependencies": {
+    "turbo": "^1.1.6"
+  },
   "scripts": {
+    "build": "turbo run build",
+    "dev": "turbo run dev --no-cache --parallel --continue",
+    "storybook":"turbo run storybook --no-cache --parallel --continue",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [
