+++ pnpm-workspace.yaml
@@ -1,0 +1,5 @@
+packages:
+  - "lib/**"
+  # exclude packages that are inside test directories
+  - "!**/test/**"
+  - "!**/__tests__/**"
