+++ turbo.json
@@ -1,0 +1,15 @@
+{
+  "baseBranch": "origin/main",
+  "pipeline": {
+    "build": {
+      "dependsOn": ["^build"],
+      "outputs": ["dist/**", "src/**"]
+    },
+    "dev": {
+      "cache": false
+    },
+    "storybook": {
+      "cache": false
+    }
+  }
+}
